feat: hydration utilities moved to main package, renamed `ssr` to `nuxt`
