+++ examples/nuxt-simple/layouts/default.vue
@@ -7,7 +7,7 @@
 
 <script lang="ts">
 import { defineComponent } from "@nuxtjs/composition-api";
-import { useNuxtQueryProvider } from "vue-query/ssr";
+import { useNuxtQueryProvider } from "vue-query/nuxt";
 import { VueQueryDevTools } from "vue-query/devtools";
 
 export default defineComponent({
