+++ package.json
@@ -7,7 +7,7 @@
   "types": "lib/index.d.ts",
   "files": [
     "/devtools",
-    "/ssr",
+    "/nuxt",
     "/lib",
     "/esm",
     "/types"
