+++ src/index.ts
@@ -10,4 +10,11 @@ export {
   MutationCache,
 } from "react-query/core";
 
+export { hydrate, dehydrate } from "./hydration";
+export type {
+  DehydrateOptions,
+  DehydratedState,
+  HydrateOptions,
+} from "./hydration";
+
 export * from "./vue";
