+++ src/nuxt/index.ts
@@ -1,6 +1,4 @@
 /* istanbul ignore file */
 
-export { hydrate, dehydrate } from "./hydration";
-
 export { useNuxtQueryProvider } from "./useNuxtQueryProvider";
 export { useNuxtDehydrate } from "./useNuxtDehydrate";
